# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Moonlit Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStats {
  totalGuilds: number;
  totalWarnings: number;
  totalBannedWords: number;
  totalCustomCommands: number;
  totalQueueEntries: number;
}

export interface GuildSummary {
  guildId: string;
  warningCount: number;
  queueCount: number;
  bannedWordCount: number;
  customCommandCount: number;
  automodEnabled: boolean;
}

export interface GuildSettings {
  guildId: string;
  /** @nullable */
  logChannel?: string | null;
  /** @nullable */
  welcomeChannel?: string | null;
  /** @nullable */
  welcomeMessage?: string | null;
  /** @nullable */
  leaveMessage?: string | null;
  prefix: string;
  automodEnabled: boolean;
  antiInvite: boolean;
  antiMentionSpam: boolean;
  mentionLimit: number;
}

export interface GuildSettingsInput {
  /** @nullable */
  logChannel?: string | null;
  /** @nullable */
  welcomeChannel?: string | null;
  /** @nullable */
  welcomeMessage?: string | null;
  /** @nullable */
  leaveMessage?: string | null;
  prefix?: string;
  automodEnabled?: boolean;
  antiInvite?: boolean;
  antiMentionSpam?: boolean;
  mentionLimit?: number;
}

export interface Warning {
  id: number;
  guildId: string;
  userId: string;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  moderatorId?: string | null;
  createdAt: string;
}

export interface UserWarningStat {
  userId: string;
  count: number;
}

export interface QueueEntry {
  userId: string;
  position: number;
  joinedAt: string;
}

export interface BannedWord {
  word: string;
}

export interface BannedWordInput {
  /** @minLength 1 */
  word: string;
}

export interface CustomCommand {
  name: string;
  response: string;
}

export interface CustomCommandInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  response: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}