# lib/api-zod/src/generated/types/bannedWordInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Moonlit Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BannedWordInput {
  /** @minLength 1 */
  word: string;
}