# lib/api-zod/src/generated/types/customCommand.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Moonlit Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomCommand {
  name: string;
  response: string;
}