# lib/api-zod/src/generated/types/customCommandInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Moonlit Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomCommandInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  response: string;
}