# lib/api-zod/src/generated/types/guildSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Moonlit Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GuildSettings {
  guildId: string;
  /** @nullable */
  logChannel?: string | null;
  /** @nullable */
  welcomeChannel?: string | null;
  /** @nullable */
  welcomeMessage?: string | null;
  /** @nullable */
  leaveMessage?: string | null;
  prefix: string;
  automodEnabled: boolean;
  antiInvite: boolean;
  antiMentionSpam: boolean;
  mentionLimit: number;
}