# lib/api-zod/src/generated/types/warning.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Moonlit Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Warning {
  id: number;
  guildId: string;
  userId: string;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  moderatorId?: string | null;
  createdAt: string;
}